Moves part of the PnP linker logic into an abstract class
