+++ packages/plugin-pnp/sources/PnpLinker.ts
@@ -1,12 +1,12 @@
-import {Installer, Linker, LinkOptions, MinimalLinkOptions, Manifest, LinkType, MessageName, DependencyMeta} from '@yarnpkg/core';
-import {FetchResult, Descriptor, Ident, Locator, Package, BuildDirective, BuildType}                         from '@yarnpkg/core';
-import {miscUtils, structUtils}                                                                              from '@yarnpkg/core';
-import {CwdFS, FakeFS, PortablePath, npath, ppath, toFilename, xfs}                                          from '@yarnpkg/fslib';
-import {PackageRegistry, generateInlinedScript, generateSplitScript, PnpSettings}                            from '@yarnpkg/pnp';
-import {UsageError}                                                                                          from 'clipanion';
-import mm                                                                                                    from 'micromatch';
+import {Linker, LinkOptions, MinimalLinkOptions, Manifest, MessageName, DependencyMeta} from '@yarnpkg/core';
+import {FetchResult, Ident, Locator, Package, BuildDirective, BuildType}                from '@yarnpkg/core';
+import {miscUtils, structUtils}                                                         from '@yarnpkg/core';
+import {CwdFS, FakeFS, PortablePath, npath, ppath, toFilename, xfs}                     from '@yarnpkg/fslib';
+import {generateInlinedScript, generateSplitScript, PnpSettings}                        from '@yarnpkg/pnp';
+import {UsageError}                                                                     from 'clipanion';
 
-import {getPnpPath}                                                                                          from './index';
+import {AbstractPnpInstaller}                                                           from './AbstractPnpInstaller';
+import {getPnpPath}                                                                     from './index';
 
 const FORCED_UNPLUG_PACKAGES = new Set([
   // Some packages do weird stuff and MUST be unplugged. I don't like them.
@@ -42,7 +42,7 @@ export class PnpLinker implements Linker {
   async findPackageLocator(location: PortablePath, opts: LinkOptions) {
     const pnpPath = getPnpPath(opts.project);
     if (!xfs.existsSync(pnpPath))
-      throw new UsageError(`The project in ${opts.project.cwd}/package.json doesn't seem to have been installed - running an install there might help`);
+      return null;
 
     const physicalPath = npath.fromPortablePath(pnpPath);
     const pnpFile = miscUtils.dynamicRequire(physicalPath);
@@ -60,94 +60,37 @@ export class PnpLinker implements Linker {
   }
 }
 
-class PnpInstaller implements Installer {
-  private readonly packageRegistry: PackageRegistry = new Map();
-
+class PnpInstaller extends AbstractPnpInstaller {
   private readonly unpluggedPaths: Set<string> = new Set();
-  private readonly blacklistedPaths: Set<string> = new Set();
 
-  private readonly opts: LinkOptions;
+  async getBuildScripts(locator: Locator, fetchResult: FetchResult): Promise<Array<BuildDirective>> {
+    if (!await fetchResult.packageFs.existsPromise(ppath.resolve(fetchResult.prefixPath, toFilename(`package.json`))))
+      return [];
 
-  constructor(opts: LinkOptions) {
-    this.opts = opts;
+    const buildScripts: Array<BuildDirective> = [];
+    const {scripts} = await Manifest.find(fetchResult.prefixPath, {baseFs: fetchResult.packageFs});
+
+    for (const scriptName of [`preinstall`, `install`, `postinstall`])
+      if (scripts.has(scriptName))
+        buildScripts.push([BuildType.SCRIPT, scriptName]);
+
+    // Detect cases where a package has a binding.gyp but no install script
+    const bindingFilePath = ppath.resolve(fetchResult.prefixPath, toFilename(`binding.gyp`));
+    if (!scripts.has(`install`) && fetchResult.packageFs.existsSync(bindingFilePath))
+      buildScripts.push([BuildType.SHELLCODE, `node-gyp rebuild`]);
+
+    return buildScripts;
   }
 
-  async installPackage(pkg: Package, fetchResult: FetchResult) {
-    const key1 = structUtils.requirableIdent(pkg);
-    const key2 = pkg.reference;
-
-    const buildScripts = await this.getBuildScripts(fetchResult);
-
-    if (buildScripts.length > 0 && !this.opts.project.configuration.get(`enableScripts`)) {
-      this.opts.report.reportWarningOnce(MessageName.DISABLED_BUILD_SCRIPTS, `${structUtils.prettyLocator(this.opts.project.configuration, pkg)} lists build scripts, but all build scripts have been disabled.`);
-      buildScripts.length = 0;
-    }
-
-    if (buildScripts.length > 0 && pkg.linkType !== LinkType.HARD && !this.opts.project.tryWorkspaceByLocator(pkg)) {
-      this.opts.report.reportWarningOnce(MessageName.SOFT_LINK_BUILD, `${structUtils.prettyLocator(this.opts.project.configuration, pkg)} lists build scripts, but is referenced through a soft link. Soft links don't support build scripts, so they'll be ignored.`);
-      buildScripts.length = 0;
-    }
-
-    const dependencyMeta = this.opts.project.getDependencyMeta(pkg, pkg.version);
-
-    if (buildScripts.length > 0 && dependencyMeta && dependencyMeta.built === false) {
-      this.opts.report.reportInfoOnce(MessageName.BUILD_DISABLED, `${structUtils.prettyLocator(this.opts.project.configuration, pkg)} lists build scripts, but its build has been explicitly disabled through configuration.`);
-      buildScripts.length = 0;
-    }
-
-    const hasVirtualInstances =
-      pkg.peerDependencies.size > 0 &&
-      !structUtils.isVirtualLocator(pkg) &&
-      !this.opts.project.tryWorkspaceByLocator(pkg);
-
-    const packageFs = !hasVirtualInstances && pkg.linkType !== LinkType.SOFT && (buildScripts.length > 0 || this.isUnplugged(pkg, dependencyMeta))
-      ? await this.unplugPackage(pkg, fetchResult.packageFs)
-      : fetchResult.packageFs;
-
-    const packageRawLocation = ppath.resolve(packageFs.getRealPath(), ppath.relative(PortablePath.root, fetchResult.prefixPath));
-
-    const packageLocation = this.normalizeDirectoryPath(packageRawLocation);
-    const packageDependencies = new Map<string, string | [string, string] | null>();
-    const packagePeers = new Set<string>();
-
-    for (const descriptor of pkg.peerDependencies.values()) {
-      packageDependencies.set(structUtils.requirableIdent(descriptor), null);
-      packagePeers.add(descriptor.name);
-    }
-
-    const packageStore = this.getPackageStore(key1);
-    packageStore.set(key2, {packageLocation, packageDependencies, packagePeers, linkType: pkg.linkType});
-
-    if (hasVirtualInstances)
-      this.blacklistedPaths.add(packageLocation);
-
-    return {
-      packageLocation: packageRawLocation,
-      buildDirective: buildScripts.length > 0 ? buildScripts as BuildDirective[] : null,
-    };
-  }
-
-  async attachInternalDependencies(locator: Locator, dependencies: Array<[Descriptor, Locator]>) {
-    const packageInformation = this.getPackageInformation(locator);
-
-    for (const [descriptor, locator] of dependencies) {
-      const target = !structUtils.areIdentsEqual(descriptor, locator)
-        ? [structUtils.requirableIdent(locator), locator.reference] as [string, string]
-        : locator.reference;
-
-      packageInformation.packageDependencies.set(structUtils.requirableIdent(descriptor), target);
+  async transformPackage(locator: Locator, dependencyMeta: DependencyMeta, packageFs: FakeFS<PortablePath>, {hasBuildScripts}: {hasBuildScripts: boolean}) {
+    if (hasBuildScripts || this.isUnplugged(locator, dependencyMeta)) {
+      return this.unplugPackage(locator, packageFs);
+    } else {
+      return packageFs;
     }
   }
 
-  async attachExternalDependents(locator: Locator, dependentPaths: Array<PortablePath>) {
-    for (const dependentPath of dependentPaths) {
-      const packageInformation = this.getDiskInformation(dependentPath);
-
-      packageInformation.packageDependencies.set(structUtils.requirableIdent(locator), locator.reference);
-    }
-  }
-
-  async finalizeInstall() {
+  async finalizeInstallWithPnp(pnpSettings: PnpSettings) {
     if (this.opts.project.configuration.get(`nodeLinker`) !== `pnp`)
       return;
 
@@ -159,52 +102,8 @@ class PnpInstaller implements Installer {
       }
     }
 
-    this.trimBlacklistedPackages();
-
-    this.packageRegistry.set(null, new Map([
-      [null, this.getPackageInformation(this.opts.project.topLevelWorkspace.anchoredLocator)],
-    ]));
-
-    const buildIgnorePattern = (ignorePatterns: Array<string>) => {
-      if (ignorePatterns.length === 0)
-        return null;
-
-
-      return ignorePatterns.map(pattern => {
-        return `(${mm.makeRe(pattern, {
-          // @ts-ignore
-          windows: false,
-        }).source})`;
-      }).join(`|`);
-    };
-
-    const pnpFallbackMode = this.opts.project.configuration.get(`pnpFallbackMode`);
-
-    const blacklistedLocations = this.blacklistedPaths;
-    const dependencyTreeRoots = this.opts.project.workspaces.map(({anchoredLocator}) => ({name: structUtils.requirableIdent(anchoredLocator), reference: anchoredLocator.reference}));
-    const enableTopLevelFallback = pnpFallbackMode !== `none`;
-    const fallbackExclusionList = [];
-    const ignorePattern = buildIgnorePattern([`.vscode/pnpify/**`, ...this.opts.project.configuration.get(`pnpIgnorePatterns`)]);
-    const packageRegistry = this.packageRegistry;
-    const shebang = this.opts.project.configuration.get(`pnpShebang`);
-
-    if (pnpFallbackMode === `dependencies-only`)
-      for (const pkg of this.opts.project.storedPackages.values())
-        if (this.opts.project.tryWorkspaceByLocator(pkg))
-          fallbackExclusionList.push({name: structUtils.requirableIdent(pkg), reference: pkg.reference});
-
     const pnpPath = getPnpPath(this.opts.project);
     const pnpDataPath = this.opts.project.configuration.get(`pnpDataPath`);
-
-    const pnpSettings: PnpSettings = {
-      blacklistedLocations,
-      dependencyTreeRoots,
-      enableTopLevelFallback,
-      fallbackExclusionList,
-      ignorePattern,
-      packageRegistry,
-      shebang,
-    };
 
     if (this.opts.project.configuration.get(`pnpEnableInlining`)) {
       const loaderFile = generateInlinedScript(pnpSettings);
@@ -237,58 +136,6 @@ class PnpInstaller implements Installer {
     }
   }
 
-  private getPackageStore(key: string) {
-    let packageStore = this.packageRegistry.get(key);
-
-    if (!packageStore)
-      this.packageRegistry.set(key, packageStore = new Map());
-
-    return packageStore;
-  }
-
-  private getPackageInformation(locator: Locator) {
-    const key1 = structUtils.requirableIdent(locator);
-    const key2 = locator.reference;
-
-    const packageInformationStore = this.packageRegistry.get(key1);
-    if (!packageInformationStore)
-      throw new Error(`Assertion failed: The package information store should have been available (for ${structUtils.prettyIdent(this.opts.project.configuration, locator)})`);
-
-    const packageInformation = packageInformationStore.get(key2);
-    if (!packageInformation)
-      throw new Error(`Assertion failed: The package information should have been available (for ${structUtils.prettyLocator(this.opts.project.configuration, locator)})`);
-
-    return packageInformation;
-  }
-
-  private getDiskInformation(path: PortablePath) {
-    const packageStore = this.getPackageStore(`@@disk`);
-    const normalizedPath = this.normalizeDirectoryPath(path);
-
-    let diskInformation = packageStore.get(normalizedPath);
-
-    if (!diskInformation) {
-      packageStore.set(normalizedPath, diskInformation = {
-        packageLocation: normalizedPath,
-        packageDependencies: new Map(),
-        packagePeers: new Set(),
-        linkType: LinkType.SOFT,
-      });
-    }
-
-    return diskInformation;
-  }
-
-  private trimBlacklistedPackages() {
-    for (const packageStore of this.packageRegistry.values()) {
-      for (const [key2, packageInformation] of packageStore) {
-        if (this.blacklistedPaths.has(packageInformation.packageLocation)) {
-          packageStore.delete(key2);
-        }
-      }
-    }
-  }
-
   private async locateNodeModules() {
     const nodeModules: Array<PortablePath> = [];
 
@@ -305,35 +152,6 @@ class PnpInstaller implements Installer {
     }
 
     return nodeModules;
-  }
-
-  private normalizeDirectoryPath(folder: PortablePath) {
-    let relativeFolder = ppath.relative(this.opts.project.cwd, folder);
-
-    if (!relativeFolder.match(/^\.{0,2}\//))
-      // Don't use ppath.join here, it ignores the `.`
-      relativeFolder = `./${relativeFolder}` as PortablePath;
-
-    return relativeFolder.replace(/\/?$/, '/')  as PortablePath;
-  }
-
-  private async getBuildScripts(fetchResult: FetchResult) {
-    if (!await fetchResult.packageFs.existsPromise(ppath.resolve(fetchResult.prefixPath, toFilename(`package.json`))))
-      return [];
-
-    const buildScripts = [];
-    const {scripts} = await Manifest.find(fetchResult.prefixPath, {baseFs: fetchResult.packageFs});
-
-    for (const scriptName of [`preinstall`, `install`, `postinstall`])
-      if (scripts.has(scriptName))
-        buildScripts.push([BuildType.SCRIPT, scriptName]);
-
-    // Detect cases where a package has a binding.gyp but no install script
-    const bindingFilePath = ppath.resolve(fetchResult.prefixPath, toFilename(`binding.gyp`));
-    if (!scripts.has(`install`) && fetchResult.packageFs.existsSync(bindingFilePath))
-      buildScripts.push([BuildType.SHELLCODE, `node-gyp rebuild`]);
-
-    return buildScripts;
   }
 
   private getUnpluggedPath(locator: Locator) {
